pyproject.toml
[project]
name = "comfyui-naia-bridge"
version = "0.1.0"
description = "ComfyUI custom node package for bridging NAIA 2.0 prompts into ComfyUI workflows via its Remote Web API."
authors = [
    { name = "DNT-LAB", email = "meno9428@gmail.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["comfyui", "comfyui-nodes", "naia", "nai", "novelai", "prompt"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "websocket-client>=1.5",
    "requests>=2.28",
]

[project.urls]
Repository = "https://github.com/DNT-LAB/comfyui-naia-bridge"
Issues = "https://github.com/DNT-LAB/comfyui-naia-bridge/issues"

[tool.comfy]
PublisherId = "DNT-LAB"
DisplayName = "ComfyUI NAIA 2.0 Bridge"
# Icon = ""  # Registry에서 나중에 업로드 가능. 400x400 이하 PNG 권장.
# requires-comfyui = ">=0.3.0"
